packages/tta-dev-observability/pyproject.toml
[project]
name = "tta-dev-observability"
version = "0.1.0"
description = "Observability and tracing extensions for TTA development primitives"
authors = [{ name = "TTA Development Team" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "tta-dev-primitives>=0.1.0",
]

[project.optional-dependencies]
tracing = [
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-instrumentation>=0.45b0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tta_dev_observability"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "UP", "ANN"]
ignore = ["E501", "ANN101", "ANN102"]

[tool.uv.sources]
tta-dev-primitives = { path = "../tta-dev-primitives", editable = true }
